Add files via upload
Mike O'Dea and I worked together and bounced ideas off each other, but we weren't sure if we could work in a group or not. I'm only noting this because I'm not sure if he submitted the answers with both our names on it or if he only included his name. Sorry if there is any confusion with this.